% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_chains.R
\name{compare_chains}
\alias{compare_chains}
\title{Compare chains of posterior trees}
\usage{
compare_chains(
  ...,
  f_tree = patristic,
  f_chain = euclidean,
  adonis2_args = list()
)
}
\arguments{
\item{...}{Two or more posterior chains of transmission trees. Each chain is a list containing one or more data frames,
where each data frame represents a transmission tree with columns \code{from} (infector) and \code{to} (infectee).}

\item{f_tree}{A function to compute the pairwise distances between nodes within a tree.
Must take a single tree (data frame) as input and return a square matrix where dimensions match the number of nodes.
Default is \code{patristic}.}

\item{f_chain}{A function to compute the distance between two chains of transmission trees.
It must take two square matrices (produced by \code{f_tree}) as input and return a single numeric distance value.
Default is \code{euclidean}.}

\item{adonis2_args}{A list of additional arguments to pass to \code{vegan::adonis2}. Default is an empty list.}
}
\value{
A \code{vegan::adonis2} object containing the results of the PERMANOVA.
}
\description{
Performs a PERMANOVA (adonis2) to test for significant differences between posterior chains of transmission trees.
}
