% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_dist.R
\name{euclidean}
\alias{euclidean}
\title{Calculate the Euclidean distance between two distance matrices.}
\usage{
euclidean(mat1, mat2)
}
\arguments{
\item{mat1}{A numeric matrix.}

\item{mat2}{A numeric matrix.}
}
\value{
A numeric value representing the Euclidean distance between the
lower triangular parts of \code{mat1} and \code{mat2}.
}
\description{
This function computes the Euclidean distance between the lower triangular
parts of two given matrices.
}
\examples{
mat1 <- matrix(c(1, 2, 3, 4), 2, 2)
mat2 <- matrix(c(4, 3, 2, 1), 2, 2)
euclidean(mat1, mat2)
}
