% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_dist.R
\name{patristic}
\alias{patristic}
\title{Compute the Patristic distance matrix}
\usage{
patristic(tree)
}
\arguments{
\item{tree}{A data frame representing a transmission tree, with the first column containing the infector IDs and the second the infectee IDs.}
}
\value{
A square, symmetric matrix of patristic distances between nodes.
}
\description{
The patristic distance is the number of generations separating any two nodes in a transmission tree.
}
\examples{
tree <- data.frame(from = c(1, 1, 2, 2, 3, 3), to = c(2, 3, 4, 5, 6, 7))
patristic(tree)
}
