% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tree.R
\name{process_tree}
\alias{process_tree}
\title{Process a Transmission Tree}
\usage{
process_tree(tree, silent = FALSE)
}
\arguments{
\item{tree}{A data frame with columns \code{from} and \code{to}.}
}
\value{
The validated and potentially cleaned tree.
}
\description{
Will attempt to validate the tree.
If the tree is invalid, it will remove the introduction and try again.
}
\keyword{internal}
