% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tree.R
\name{remove_intro}
\alias{remove_intro}
\title{Remove the Introduction from a Transmission Tree}
\usage{
remove_intro(tree)
}
\arguments{
\item{tree}{A data frame with columns \code{from} and \code{to} representing the transmission tree.}
}
\value{
A cleaned tree with the introduction removed.
}
\description{
Will remove the single introduction (row with an \code{NA} in the \code{from} column).
}
\keyword{internal}
