% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tree.R
\name{validate_set}
\alias{validate_set}
\title{Validate a Set of Transmission Trees}
\usage{
validate_set(set)
}
\arguments{
\item{set}{A list containing at least one dataframe.}
}
\value{
Invisible \code{TRUE} if the set is valid. Throws an error if invalid.
}
\description{
Ensures that the input is a list containing at least one dataframe.
}
\keyword{internal}
